requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_metadata==8.6.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
gunicorn==23.0.0